+++ src/Snaptcha.php
@@ -104,6 +104,7 @@ class Snaptcha extends Plugin
             || $request->getIsPreview()
             || $request->getMethod() !== 'POST'
             || $request->getFullPath() == Craft::$app->getConfig()->getGeneral()->getSetPasswordPath()
+            || $this->snaptcha->isExcludedControllerAction()
             || $this->snaptcha->isExcludedUri($request->getUrl())
         ) {
             return;